==> empresa4/02_make_datasets_clase_2.ipynb <==
import pandas as pd
import numpy as np
from core import filter_clientes_importantes, filter_productos_importantes

# Reading the CSV file
original_data = pd.read_csv("./datasets/originales_clase_3/tb_sellout_02.txt.gz", sep="\t")
original_categories = pd.read_csv("./datasets/originales_clase_3/tb_productos_02.txt", sep="\t")

original_categories

original_categories["product_id"].nunique()

# drop duplicate product_id from the original_categories dataframe, keep first row
original_categories.drop_duplicates(subset="product_id", keep="first", inplace=True)
original_categories["product_id"].nunique()


original_data["periodo"].value_counts().sort_index()

original_categories.to_csv("./datasets/maestro_productos.csv", index=False)

original_data["product_id"].nunique(), original_data["customer_id"].nunique(), original_data["periodo"].nunique() 

df = original_data

df[df["plan_precios_cuidados"] == 1]

df[(df["product_id"] == 20405)&(df["periodo"] > 201704)].sort_values(by="periodo")

df['periodo'] = df['periodo'].astype(int)
df['product_id'] = df['product_id'].astype(int)
df['customer_id'] = df['customer_id'].astype(int)
df['plan_precios_cuidados'] = df['plan_precios_cuidados'].astype(int)
df['cust_request_qty'] = df['cust_request_qty'].astype(float)
df['cust_request_tn'] = df['cust_request_tn'].astype(float)
df['tn'] = df['tn'].astype(float)

print(df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())


# print unique number of customer_id, product_id, and period
print(df["customer_id"].nunique(), df["product_id"].nunique(), df["periodo"].nunique())

df_clientes = df.groupby(['periodo', 'customer_id', 'product_id'])[['cust_request_qty', 'cust_request_tn', 'tn']].sum()
df_productos = df.groupby(['periodo', 'product_id', ])[['cust_request_qty', 'cust_request_tn', 'tn']].sum()

print("ORIGINAL", df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())
print("CLIENTES", df_clientes["cust_request_qty"].sum(), df_clientes["cust_request_tn"].sum(), df_clientes["tn"].sum())
print("PRODUCTOS", df_productos["cust_request_qty"].sum(), df_productos["cust_request_tn"].sum(), df_productos["tn"].sum())


df_productos.reset_index(inplace=True)
df_clientes.reset_index(inplace=True)

product_ids_maestro = original_categories["product_id"].unique().tolist()
product_ids_sellout = df_productos["product_id"].unique().tolist()
faltantes_en_sellout = list(set(product_ids_maestro) - set(product_ids_sellout))
faltantes_en_maestro = list(set(product_ids_sellout) - set(product_ids_maestro))
len(faltantes_en_sellout), len(faltantes_en_maestro)

# for each value in faltantes_en_sellout, add a row to df_productos with 0 values and periodo = 201701
for product_id in faltantes_en_sellout:
    df_productos = pd.concat([df_productos, pd.DataFrame([{'product_id': product_id, 'periodo': 201701, 'cust_request_qty': 0, 'cust_request_tn': 0, 'tn': 0}])], axis=0, ignore_index=True)
    df_clientes = pd.concat([df_clientes, pd.DataFrame([{'product_id': product_id, 'periodo': 201701, 'customer_id': 10001, 'cust_request_qty': 0, 'cust_request_tn': 0, 'tn': 0}])], axis=0, ignore_index=True)
df_productos["product_id"].nunique(), df_clientes["product_id"].nunique()

df_clientes


# Find all unique values for 'periodo' and 'product_id'
unique_periodos = df['periodo'].unique()
unique_product_ids = df_productos['product_id'].unique()
unique_customer_ids = df['customer_id'].unique()
len(unique_periodos), len(unique_product_ids), len(unique_customer_ids)

print("Shape antes: ", df_productos.shape)

from itertools import product

# Create all possible combinations of 'periodo' and 'product_id'
all_combinations = list(product(unique_periodos, unique_product_ids))

# Convert the combinations into a DataFrame
all_combinations_df = pd.DataFrame(all_combinations, columns=['periodo', 'product_id'])

# Merge with the existing DataFrame to find missing rows
combined_df = all_combinations_df.merge(df_productos, on=['periodo', 'product_id'], how='left')

# Identify the missing rows by looking for NaNs
missing_rows = combined_df[combined_df['cust_request_qty'].isna()]

# For the missing rows, set 'cust_request_qty', 'cust_request_tn', and 'tn' to 0
# We also need to decide how to handle 'product_category'. For now, we'll set it as 'Unknown'
missing_rows.loc[:, 'cust_request_qty'] = 0
missing_rows.loc[:, 'cust_request_tn'] = 0
missing_rows.loc[:, 'tn'] = 0

# Now we append the missing rows to the original DataFrame
# convert to pd.concat: complete_df = df_productos.append(missing_rows[df_productos.columns], ignore_index=True)
df_productos = pd.concat([df_productos, missing_rows[df_productos.columns]], ignore_index=True)

# Return the number of new rows added and the path to the new file
num_new_rows = len(missing_rows)
print(num_new_rows)

print("Shape Después: ", df_productos.shape)


# Debería tener resto 0
print(df_productos.shape[0] / 36)

print("Shape antes: ", df_clientes.shape)
print(df_clientes.shape[0] / 36)

from itertools import product

# Create all possible combinations of 'periodo' and 'product_id'
all_combinations = list(product(unique_periodos, unique_customer_ids, unique_product_ids))

# Convert the combinations into a DataFrame
all_combinations_df = pd.DataFrame(all_combinations, columns=['periodo', 'customer_id', 'product_id'])

# Merge with the existing DataFrame to find missing rows
combined_df = all_combinations_df.merge(df_clientes, on=['periodo', 'customer_id', 'product_id'], how='left')

# Identify the missing rows by looking for NaNs
missing_rows = combined_df[combined_df['cust_request_qty'].isna()]

# For the missing rows, set 'cust_request_qty', 'cust_request_tn', and 'tn' to 0
# We also need to decide how to handle 'product_category'. For now, we'll set it as 'Unknown'
missing_rows.loc[:, 'cust_request_qty'] = 0
missing_rows.loc[:, 'cust_request_tn'] = 0
missing_rows.loc[:, 'tn'] = 0

# Now we append the missing rows to the original DataFrame
# convert to pd.concat: complete_df = df_productos.append(missing_rows[df_productos.columns], ignore_index=True)
df_clientes = pd.concat([df_clientes, missing_rows[df_clientes.columns]], ignore_index=True)

# Return the number of new rows added and the path to the new file
num_new_rows = len(missing_rows)
print(num_new_rows)

print("Shape Después: ", df_clientes.shape)

# Debería tener resto 0
print(df_clientes.shape[0] / 36)


df_productos = pd.merge(df_productos, original_categories[['product_id', 'cat1', 'cat2', 'sku_size']], on='product_id', how='left')
df_productos.rename(columns={'cat1': 'product_category'}, inplace=True)
df_productos["product_category"].fillna('unknown', inplace=True)
df_productos["cat2"].fillna('unknown', inplace=True)
df_productos["sku_size"].fillna(0, inplace=True)

df_clientes = pd.merge(df_clientes, original_categories[['product_id', 'cat1', 'cat2', 'sku_size']], on='product_id', how='left')
df_clientes.rename(columns={'cat1': 'product_category'}, inplace=True)
df_clientes["product_category"].fillna('unknown', inplace=True)  
df_clientes["cat2"].fillna('unknown', inplace=True)
df_clientes["sku_size"].fillna(0, inplace=True)

print("ORIGINAL", df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())
print("CLIENTES", df_clientes["cust_request_qty"].sum(), df_clientes["cust_request_tn"].sum(), df_clientes["tn"].sum())
print("PRODUCTOS", df_productos["cust_request_qty"].sum(), df_productos["cust_request_tn"].sum(), df_productos["tn"].sum())

df_precios_cuidados = df[df["plan_precios_cuidados"] == 1]
# find rows with duplicated combinations of 'periodo' and 'product_id' and remove them
df_precios_cuidados = df_precios_cuidados.drop_duplicates(subset=['periodo', 'product_id'], keep='first')
df_precios_cuidados["product_id"].nunique(), df_precios_cuidados.shape

df_precios_cuidados = df_precios_cuidados[['periodo', 'product_id', 'plan_precios_cuidados']]
df_precios_cuidados

# add a column "plan_precios_cuidados" to df_productos based on the values in df_precios_cuidados
df_productos = pd.merge(df_productos, df_precios_cuidados, on=['periodo', 'product_id'], how='left')
df_productos.fillna(0, inplace=True)

df_clientes = pd.merge(df_clientes, df_precios_cuidados, on=['periodo', 'product_id'], how='left')
df_clientes.fillna(0, inplace=True)

# convert plan_precios_cuidados to int
df_productos['plan_precios_cuidados'] = df_productos['plan_precios_cuidados'].astype(int)
df_clientes['plan_precios_cuidados'] = df_clientes['plan_precios_cuidados'].astype(int)
df_productos["plan_precios_cuidados"].value_counts(), df_clientes["plan_precios_cuidados"].value_counts()  

df_producto_estrella = df_productos[df_productos['product_id'] == 20001]


df_producto_estrella


# show rows with "unknown" in any column
df_productos[df_productos.isin(['unknown']).any(axis=1)]

df_producto_estrella.isna().sum()

df_clientes.isna().sum()

df_120 = filter_productos_importantes(df_clientes)
df_120 = filter_clientes_importantes(df_120)
df_120.shape[0] / 36

df.to_csv('./datasets/tb_sellout_02_original.csv', index=False)
df_precios_cuidados.to_csv('./datasets/tb_sellout_02_precios_cuidados.csv', index=False)
df_producto_estrella.to_csv('./datasets/tb_sellout_02_producto_estrella.csv', index=False)
df_productos.to_csv('./datasets/tb_sellout_02_productos_todos.csv', index=False)
df_120.to_csv('./datasets/tb_sellout_02_120.csv', index=False)

df_productos

df_productos_anti_leak = df_productos[df_productos["periodo"] <= 201902]
df_productos_anti_leak.to_csv('./datasets/tb_sellout_02_productos_todos_anti_leak.csv', index=False)

df_clientes

df_120

df_120_anti_leak = df_120[df_120["periodo"] <= 201902]
df_120_anti_leak.to_csv('./datasets/tb_sellout_02_120_anti_leak.csv', index=False)

df_120_anti_leak
