==> empresa4/02_make_datasets_clase_2_stock.ipynb <==
import pandas as pd
import numpy as np
from core import filter_clientes_importantes, filter_productos_importantes

# Reading the CSV file
stocks_02 = pd.read_csv("./datasets/originales_clase_3/tb_stocks_02.txt.gz", sep="\t")
stocks_01 = pd.read_csv("./datasets/originales_clase_2/tb_stocks_01.txt.gz", sep="\t", header=None)
productos_todos = pd.read_csv("./datasets/tb_sellout_02_productos_todos.csv")
maestro = pd.read_csv("./datasets/maestro_productos.csv")

stocks_01

columns = 3
data_series = stocks_01.iloc[:, 0]
column_names = data_series[:columns].tolist()
data_series = data_series[columns:]

# Ensuring the data length is a multiple of 5
data_series = data_series[: len(data_series) - (len(data_series) % columns)]

# Reshaping the data
reshaped_data = data_series.values.reshape(-1, columns)

stocks_01 = pd.DataFrame(
    reshaped_data,
    columns=column_names,
)
stocks_01

stocks_01["periodo"].value_counts()

stocks_02["periodo"].value_counts()

# group original_data by periodo and product_id, sum stock_final
stocks_02 = stocks_02.groupby(['periodo', 'product_id']).agg({'stock_final': 'sum'}).reset_index()
stocks_02

product_ids_maestro = maestro["product_id"].unique().tolist()
product_ids_sellout = productos_todos["product_id"].unique().tolist()
product_ids_stock = stocks_02["product_id"].unique().tolist()
faltantes_en_stock = list(set([*product_ids_sellout, *product_ids_maestro]) - set(product_ids_stock))
len(faltantes_en_stock)

periodos = productos_todos["periodo"].unique().tolist()
periodos_stock_02 = stocks_02["periodo"].unique().tolist()
periodos_faltantes = list(set(periodos) - set(periodos_stock_02))
len(periodos_faltantes)

# for each value in faltantes_en_stock, add a row to original_data with 0 values and periodo = 201701
for product_id in faltantes_en_stock:
    stocks_02 = pd.concat([stocks_02, pd.DataFrame([{'product_id': product_id, 'periodo': 201701, 'stock_final': 0}])], axis=0, ignore_index=True)
stocks_02["product_id"].nunique()

# for each value in periodos, add a row to original_data with 0 values and periodo = 201701
for periodo in periodos_faltantes:
    stocks_02 = pd.concat([stocks_02, pd.DataFrame([{'product_id': 20001, 'periodo': periodo, 'stock_final': 0}])], axis=0, ignore_index=True)
stocks_02["periodo"].nunique()

# Find all unique values for 'periodo' and 'product_id'
unique_periodos = stocks_02['periodo'].unique()
unique_product_ids = stocks_02['product_id'].unique()
len(unique_periodos), len(unique_product_ids)

print("Shape antes: ", stocks_02.shape)

from itertools import product

# Create all possible combinations of 'periodo' and 'product_id'
all_combinations = list(product(unique_periodos, unique_product_ids))

# Convert the combinations into a DataFrame
all_combinations_df = pd.DataFrame(all_combinations, columns=['periodo', 'product_id'])

# Merge with the existing DataFrame to find missing rows
combined_df = all_combinations_df.merge(stocks_02, on=['periodo', 'product_id'], how='left')

# Identify the missing rows by looking for NaNs
missing_rows = combined_df[combined_df['stock_final'].isna()]

missing_rows.loc[:, 'stock_final'] = 0

# Now we append the missing rows to the original DataFrame
# convert to pd.concat: complete_df = stocks_02.append(missing_rows[stocks_02.columns], ignore_index=True)
stocks_02 = pd.concat([stocks_02, missing_rows[stocks_02.columns]], ignore_index=True)

# Return the number of new rows added and the path to the new file
num_new_rows = len(missing_rows)
print(num_new_rows)

print("Shape Después: ", stocks_02.shape)


stocks_02.drop_duplicates(subset=['product_id', 'periodo'], keep='first', inplace=True)

# Debería tener resto 0
print(stocks_02.shape[0] / 36)

stocks_02.isna().sum()

print("ORIGINAL", pd.read_csv("./datasets/originales_clase_3/tb_stocks_02.txt.gz", sep="\t")["stock_final"].sum())
print("PRODUCTOS", stocks_02["stock_final"].sum())

stocks_02.to_csv('./datasets/tb_stocks_02_productos_todos.csv', index=False)

stocks_02 = stocks_02[stocks_02["periodo"] <= 201902]
stocks_02.to_csv('./datasets/tb_stocks_02_productos_todos_anti_leak.csv', index=False)

# Debería tener resto 0
print(stocks_02.shape[0] / 26)
