==> empresa4/03_make_datasets_clase_2_desglosado.ipynb <==
import pandas as pd
import numpy as np
from core import filter_clientes_importantes, filter_productos_importantes

# 1. Crudo Desglosado

## Cargamos originales

# Reading the CSV file
original_data = pd.read_csv("./datasets/originales_clase_3/tb_sellout_02.txt.gz", sep="\t")
original_categories = pd.read_csv("./datasets/originales_clase_3/tb_productos_02.txt", sep="\t")

original_data

original_categories

unknown_category_product_ids = [20808, 20848, 21066, 21098, 21125, 21165, 21178, 21199, 21217,
       21223, 21225, 21230, 21238, 21240, 21241, 21249, 21253, 21268,
       21272, 21273, 21274, 21275, 21277, 21285, 21286, 21295, 21297,
       21281, 21278, 21284, 21279, 21283, 21289, 21290, 21288, 21291,
       21292, 21293, 21296, 21298, 21299, 20918, 21228, 21270, 21169]
original_categories[original_categories["product_id"].isin(unknown_category_product_ids)]

original_categories["product_id"].nunique()

# drop duplicate product_id from the original_categories dataframe, keep first row
original_categories.drop_duplicates(subset="product_id", keep="first", inplace=True)
original_categories["product_id"].nunique()

original_data["periodo"].value_counts().sort_index()

original_data["product_id"].nunique(), original_data["customer_id"].nunique(), original_data["periodo"].nunique() 

df = original_data

df[df["plan_precios_cuidados"] == 1]

## Limpiamos types

df['periodo'] = df['periodo'].astype(int)
df['product_id'] = df['product_id'].astype(int)
df['customer_id'] = df['customer_id'].astype(int)
df['plan_precios_cuidados'] = df['plan_precios_cuidados'].astype(int)
df['cust_request_qty'] = df['cust_request_qty'].astype(float)
df['cust_request_tn'] = df['cust_request_tn'].astype(float)
df['tn'] = df['tn'].astype(float)

print(df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())


# print unique number of customer_id, product_id, and period
print(df["product_id"].nunique(), df["customer_id"].nunique(), df["periodo"].nunique())

## Agrupamos para desagregar

df_desagregado = df.groupby(['periodo', 'customer_id', 'product_id'])[['cust_request_qty', 'cust_request_tn', 'tn']].sum()

print("ORIGINAL|||", df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())
print("DESAGREGADO", df_desagregado["cust_request_qty"].sum(), df_desagregado["cust_request_tn"].sum(), df_desagregado["tn"].sum())

df_desagregado.reset_index(inplace=True)

product_ids_maestro = original_categories["product_id"].unique().tolist()
product_ids_sellout = df_desagregado["product_id"].unique().tolist()
faltantes_en_sellout = list(set(product_ids_maestro) - set(product_ids_sellout))
faltantes_en_maestro = list(set(product_ids_sellout) - set(product_ids_maestro))
len(faltantes_en_sellout), len(faltantes_en_maestro)

df_desagregado.sort_values(by=["periodo", "customer_id", "product_id"], inplace=True)
df_desagregado

print("ORIGINAL|||", df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())
print("DESAGREGADO", df_desagregado["cust_request_qty"].sum(), df_desagregado["cust_request_tn"].sum(), df_desagregado["tn"].sum())

## Agregamos los 0s donde faltan y eliminamos productos discontinuados

# Inicializa un DataFrame vacío para almacenar los resultados
df_final = pd.DataFrame()

# Empezamos capturando únicamente los productos que existen en el primer periodo
existing_products = df_desagregado[df_desagregado["periodo"] == 201701]["product_id"].unique().tolist()

# Y capturamos todos los clientes (que no se crean, sino que siempre existen todos)
total_clientes = df_desagregado["customer_id"].unique().tolist()

missing_products_dict = {}

for periodo in df_desagregado["periodo"].unique().tolist():
    print(periodo)
    print("existing_products before", len(existing_products))
    
    # 1) get the products that exist in the current period
    this_period_products = df_desagregado[df_desagregado["periodo"] == periodo]["product_id"].unique().tolist()
    print("this_period_products", len(this_period_products))
    for product in this_period_products:
        if product in missing_products_dict:
            del missing_products_dict[product]

    # 2) get the products that are present in existing_products but not in the current period
    missing_products = list(set(existing_products) - set(this_period_products))
    print("missing_products", len(missing_products))
    for product in missing_products:
        if product in missing_products_dict:
            missing_products_dict[product] += 1
        else:
            missing_products_dict[product] = 1

    # 3) update the existing products list
    existing_products = list(set([*existing_products, *this_period_products]))
    print("existing_products after", len(existing_products))

    for product in missing_products_dict:
        if missing_products_dict[product] >= 3:
            # remove the product from the existing products list
            existing_products.remove(product) if product in existing_products else None

    # Crea todas las combinaciones posibles de clientes y productos para el periodo
    combinaciones_periodo = pd.MultiIndex.from_product([[periodo], total_clientes, existing_products], names=["periodo", "customer_id", "product_id"]).to_frame(index=False)

    # Concatena con el DataFrame final
    df_final = pd.concat([df_final, combinaciones_periodo], ignore_index=True)

# Realiza un merge con tu DataFrame original
df_merged = pd.merge(df_final, df_desagregado, on=["periodo", "customer_id", "product_id"], how='left')

# Rellena los NaN con 0 para las columnas necesarias
df_merged[['cust_request_qty', 'cust_request_tn', 'tn']] = df_merged[['cust_request_qty', 'cust_request_tn', 'tn']].fillna(0)

df_merged[df_merged["product_id"] == 20989]

missing_products_dict

df_merged

print("ORIGINAL", df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())
print("MERGED||", df_merged["cust_request_qty"].sum(), df_merged["cust_request_tn"].sum(), df_merged["tn"].sum())

# get number of duplicated rows by <period, customer_id, product_id>
df_duplicated = df_merged[df_merged.duplicated(subset=["periodo", "customer_id", "product_id"], keep=False)]
df_duplicated

df_merged.to_csv("./datasets/tb_sellout_02_desglosado_crudo.csv", index=False)

# 2. Add categories info 

df_desagregado_final = df_merged.copy()

df_desagregado_final = pd.merge(
    df_desagregado_final,
    original_categories[["product_id", "cat1", "cat2", "cat3", "brand", "sku_size"]],
    on="product_id",
    how="left",
    validate="many_to_many",
)
df_desagregado_final.rename(columns={"cat1": "product_category"}, inplace=True)
df_desagregado_final["product_category"].fillna("unknown", inplace=True)
df_desagregado_final["cat2"].fillna("unknown", inplace=True)
df_desagregado_final["cat3"].fillna("unknown", inplace=True)
df_desagregado_final["brand"].fillna("unknown", inplace=True)
df_desagregado_final["sku_size"].fillna(0, inplace=True)

print("ORIGINAL|||", df["cust_request_qty"].sum(), df["cust_request_tn"].sum(), df["tn"].sum())
print("DESAGREGADO", df_desagregado_final["cust_request_qty"].sum(), df_desagregado_final["cust_request_tn"].sum(), df_desagregado_final["tn"].sum())

df_precios_cuidados = df[df["plan_precios_cuidados"] == 1]
# find rows with duplicated combinations of 'periodo' and 'product_id' and remove them
df_precios_cuidados = df_precios_cuidados.drop_duplicates(subset=['periodo', 'product_id'], keep='first')
df_precios_cuidados["product_id"].nunique(), df_precios_cuidados.shape

df_precios_cuidados = df_precios_cuidados[['periodo', 'product_id', 'plan_precios_cuidados']]
df_precios_cuidados

# add a column "plan_precios_cuidados" to df_desagregado_final based on the values in df_precios_cuidados
df_desagregado_final = pd.merge(df_desagregado_final, df_precios_cuidados, on=['periodo', 'product_id'], how='left', validate="many_to_many")

df_desagregado_final.shape


df_desagregado_final.fillna(0, inplace=True)

# convert plan_precios_cuidados to int
df_desagregado_final['plan_precios_cuidados'] = df_desagregado_final['plan_precios_cuidados'].astype(int)
df_desagregado_final["plan_precios_cuidados"].value_counts()

# show rows with "unknown" in any column
df_desagregado_final[df_desagregado_final.isin(['unknown']).any(axis=1)]

# show rows with "unknown" in any column
df_desagregado_final[df_desagregado_final.isin(['unknown']).any(axis=1)]["product_id"].unique()

df_desagregado_final["brand"]

df_desagregado_final.shape

df_desagregado_final.to_csv("./datasets/tb_sellout_02_desglosado.csv", index=False)
