==> empresa4/04_make_time_series_dataset.ipynb <==
import pandas as pd
import requests
from io import StringIO
from dtaidistance import dtw
import numpy as np

# 1. Obtenemos la data original



url='https://drive.google.com/file/d/1xwJmomvqSsDARENHRdfI4EmIt6aXFX2x/view?usp=sharing'
file_id = url.split('/')[-2]
dwn_url='https://drive.usercontent.google.com/download?id=' + file_id + '&export=download&authuser=0&confirm=t&uuid=95f36248-0659-4956-adaa-bd1e0564383f&at=APZUnTX9peDngh8O7LsXKmRoDw12:1701132306870'
url2 = requests.get(dwn_url).text
csv_raw = StringIO(url2)
df = pd.read_csv(csv_raw)

df

# 2. Obtenemos las series de tiempo de cada <product_id, customer_id>

# make a dataframe where each row is a combination of product_id, customer_id and the columns are periodo. The value of the cells correspond to the column "tn"

df2 = df.pivot_table(index=['product_id', 'customer_id'], columns='periodo', values='tn', aggfunc='sum')
df2

df2.to_csv('time_series_dataset.csv')

# 2.B Heurísticas

time_series_df = pd.read_csv('time_series_dataset.csv', index_col=[0,1])
time_series_df

print(len(time_series_df))

# add a column with the count of columns == 0 of each row
time_series_df['count_zeros'] = time_series_df.apply(lambda x: sum(x==0), axis=1)

time_series_df[time_series_df['count_zeros'] == 36]
      

# make a new dataframe with time_series_df['count_zeros'].values_count()
df3 = pd.DataFrame(time_series_df['count_zeros'].value_counts())
df3.reset_index(inplace=True)
df3.sort_values(by='count_zeros', inplace=True)
df3

# 3. Corremos el Dynamic Time Warping para encontrar distancias

time_series = time_series_df.values
time_series

import numpy as np
from dtaidistance import dtw

def calculate_dtw_batch(time_series_batch, all_time_series):
    # Initialize an empty matrix to store DTW distances for the batch
    dtw_distances_batch = np.zeros((len(time_series_batch), len(all_time_series)))

    for i, series_i in enumerate(time_series_batch):
        for j, series_j in enumerate(all_time_series[i+1:], start=i+1):
            # Ensure series_i and series_j are 1-D arrays
            series_i_1d = np.asarray(series_i).reshape(-1)
            series_j_1d = np.asarray(series_j).reshape(-1)

            # if there are NaN at the beginning or at the end of the series, remove them
            series_i_1d = series_i_1d[~np.isnan(series_i_1d)]
            series_j_1d = series_j_1d[~np.isnan(series_j_1d)]           

            # Calculate DTW distance
            distance = dtw.distance_fast(series_i_1d, series_j_1d)
            dtw_distances_batch[i][j] = distance
            dtw_distances_batch[j][i] = distance  # Reflect the distance in the symmetric position

    return dtw_distances_batch


import time

# Sample a subset of your data
sample_size = 5000
len_dataset = len(time_series)
sample_time_series = time_series[:sample_size]  # Assuming 'time_series' is your full dataset

# Time the computation for the sample
start_time = time.time()
sample_distances = calculate_dtw_batch(sample_time_series, sample_time_series)
end_time = time.time()
time_for_sample = end_time - start_time

# Number of comparisons for the sample and the full dataset
num_comparisons_sample = sample_size * (sample_size - 1) / 2
num_comparisons_full = len_dataset * (len_dataset - 1) / 2

# Extrapolate to estimate total time
estimated_total_time = time_for_sample / num_comparisons_sample * num_comparisons_full

print(f"Time taken for sample: {time_for_sample} seconds")
print(f"Estimated total time: {estimated_total_time / 60 } minutes")
print(f"Estimated total time: {estimated_total_time / 60 / 60 } hours")
print(f"Estimated total time: {estimated_total_time / 60 / 60 / 24 } days")


num_comparisons_full

time_for_sample = 0.7444002628326416
len_dataset = len(time_series_df) / 3

# Number of comparisons for the sample and the full dataset
num_comparisons_sample = sample_size * (sample_size - 1) / 2
num_comparisons_full = len_dataset * (len_dataset - 1) / 2

# Extrapolate to estimate total time
estimated_total_time = time_for_sample / num_comparisons_sample * num_comparisons_full

print(f"Time taken for sample: {time_for_sample} seconds")
print(f"Estimated total time: {estimated_total_time / 60 } minutes")
print(f"Estimated total time: {estimated_total_time / 60 / 60 } hours")
print(f"Estimated total time: {estimated_total_time / 60 / 60 / 24 } days")

# Assume `time_series` is your data
batch_size = 10000  # Adjust the batch size based on your memory capacity
dtw_distances = []

for i in range(0, len(time_series), batch_size):
    batch = time_series[i:i + batch_size]
    dtw_distances.append(calculate_dtw_batch(batch, time_series))

# Combine the results
dtw_distances = np.vstack(dtw_distances)

# save the dtw_distances matrix to disk
np.save('dtw_distances.npy', dtw_distances)
