==> empresa4/05_make_customers_dataset_1.ipynb <==
import pandas as pd
from datasets import get_dataset, nombres_datasets

nombres_datasets

df_desglosado = get_dataset("02_productos_desglosado")
df_desglosado

# group df_desglosado by "periodo" and "customer_id", sum cust_request_qty, cust_request_tn and tn columns and discard the rest
df_desglosado_agrupado = df_desglosado.groupby(["periodo", "customer_id"]).sum()[["cust_request_qty", "cust_request_tn", "tn"]]
df_desglosado_agrupado


df_desglosado_agrupado.to_csv("./datasets/02_sellout_agrupado_por_cliente.csv")
